pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pitext-travel"
version = "1.0.0"
description = "Travel planning microservice with voice interaction"
requires-python = ">=3.8"
dependencies = [
    "Flask==2.3.2",
    "requests==2.31.0",
    "openai==1.12.0",
    "python-dotenv==1.0.0",
    "flask-socketio==5.5.1",
    "python-engineio==4.11.0",
    "python-socketio>=5.12.0",
    "numpy==1.24.3",
    "flask-cors==4.0.0",
    "tenacity==8.2.3",
    "asgiref==3.7.2"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["pitext_travel*"]